simlpify flow tcp
