+++ huginn-net-tcp/src/process/flow_state.rs
@@ -1,8 +1,7 @@
 //! Per-flow state the SYN+ACK fingerprint needs from the SYN.
 //!
-//! p0f keeps `f->syn_mss` and `f->acked` on the flow (`process.c` /
-//! `fp_tcp.c`). The peer MSS is the last window divisor tried on a response,
-//! and `acked` stops a repeated SYN+ACK from being fingerprinted twice.
+//! `syn_mss` is the last window divisor on a response. `acked` stops a
+//! repeated SYN+ACK from being fingerprinted twice.
 
 use std::net::IpAddr;
 use std::time::Duration;
@@ -13,7 +12,7 @@ const FLOW_CACHE_TTL_SECS: u64 = 30;
 
 /// Canonical client→server key for a TCP handshake.
 #[derive(Debug, Clone, Hash, Eq, PartialEq)]
-pub(crate) struct FlowKey {
+pub struct FlowKey {
     pub client_ip: IpAddr,
     pub client_port: u16,
     pub server_ip: IpAddr,
@@ -51,14 +50,14 @@ pub(crate) struct FlowState {
 
 /// What to do with a SYN+ACK once flow state has been consulted.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum SynAckDisposition {
+pub enum SynAckDisposition {
     /// A previous SYN+ACK of this flow was already fingerprinted.
     Duplicate,
     /// First SYN+ACK; carry the peer MSS the window divisors may need.
     First { peer_mss: Option<u16> },
 }
 
-pub(crate) struct FlowTracker {
+pub struct FlowTracker {
     cache: TtlCache<FlowKey, FlowState>,
 }
 
@@ -101,71 +100,5 @@ impl FlowTracker {
                 SynAckDisposition::First { peer_mss: None }
             }
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::net::{IpAddr, Ipv4Addr};
-
-    fn key() -> FlowKey {
-        FlowKey::from_syn(
-            IpAddr::V4(Ipv4Addr::new(10, 0, 0, 1)),
-            50000,
-            IpAddr::V4(Ipv4Addr::new(10, 0, 0, 2)),
-            443,
-        )
-    }
-
-    fn syn_ack_key() -> FlowKey {
-        FlowKey::from_syn_ack(
-            IpAddr::V4(Ipv4Addr::new(10, 0, 0, 2)),
-            443,
-            IpAddr::V4(Ipv4Addr::new(10, 0, 0, 1)),
-            50000,
-        )
-    }
-
-    #[test]
-    fn a_syn_ack_sees_the_mss_recorded_on_the_syn() {
-        let mut tracker = FlowTracker::new(16);
-        tracker.note_syn(key(), Some(1400));
-
-        assert_eq!(
-            tracker.begin_syn_ack(syn_ack_key()),
-            SynAckDisposition::First { peer_mss: Some(1400) }
-        );
-    }
-
-    #[test]
-    fn a_second_syn_ack_on_the_same_flow_is_a_duplicate() {
-        let mut tracker = FlowTracker::new(16);
-        tracker.note_syn(key(), Some(1400));
-        let _ = tracker.begin_syn_ack(syn_ack_key());
-
-        assert_eq!(tracker.begin_syn_ack(syn_ack_key()), SynAckDisposition::Duplicate);
-    }
-
-    #[test]
-    fn a_syn_ack_without_a_prior_syn_still_fingerprints_once() {
-        let mut tracker = FlowTracker::new(16);
-
-        assert_eq!(
-            tracker.begin_syn_ack(syn_ack_key()),
-            SynAckDisposition::First { peer_mss: None }
-        );
-        assert_eq!(tracker.begin_syn_ack(syn_ack_key()), SynAckDisposition::Duplicate);
-    }
-
-    #[test]
-    fn a_zero_mss_on_the_syn_is_not_handed_to_the_response() {
-        let mut tracker = FlowTracker::new(16);
-        tracker.note_syn(key(), Some(0));
-
-        assert_eq!(
-            tracker.begin_syn_ack(syn_ack_key()),
-            SynAckDisposition::First { peer_mss: None }
-        );
     }
 }
